fix this for everyone.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@6362 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/languages/mod_spidermonkey/mod_spidermonkey.c
@@ -2366,7 +2366,7 @@ static JSBool js_fetchurl(JSContext * cx, JSObject * obj, uintN argc, jsval * ar
 			*rval = STRING_TO_JSVAL(JS_NewStringCopyZ(cx, config_data.buffer));
 		} else {
 			char errmsg[256];
-			snprintf(errmsg, 256, "~throw new Error(\"Curl returned error %d.\");", code);
+			snprintf(errmsg, 256, "~throw new Error(\"Curl returned error %u.\");", (unsigned)code);
 			eval_some_js(errmsg, cx, obj, rval);
 		}
